+++ internal/utils/proj/proj.go
@@ -18,8 +18,8 @@ import (
 )
 
 const (
-	RadToDeg = 180 / math.Pi
-	DegToRad = math.Pi / 180
+	RadToDeg float64 = 180 / math.Pi
+	DegToRad float64 = math.Pi / 180
 )
 
 // CreateLonLatProj create a CoordinateTransform from/to the geographic lon/lat coordinates
@@ -334,6 +334,25 @@ func (s Shape) cloneTo4326(crs *godal.SpatialRef, geodetic bool) (*Shape, error)
 	return &Shape{*g}, nil
 }
 
+// mod180 returns the modulo 360 in [-180, 180]
+func mod180(f float64) float64 {
+	return math.Mod(math.Mod(f, 360)+540, 360) - 180
+}
+
+func computeDeltaOffset(lon1, lat1, lon2, lat2, lonm, latm, x1, y1, x2, y2 float64, project *godal.Transform) float64 {
+	lonm2, _ := lonLatMidPoint(lon1, lat1, lon2, lat2, true)
+	if math.Abs(lonm-lonm2) > 90 {
+		deltaLon := mod180(lonm - lon1)
+		if math.Abs(deltaLon) > 90 {
+			lonq, latq := []float64{(3*x1 + x2) / 4, (x1 + 3*x2) / 4}, []float64{(y1 + 3*y2) / 4, (3*y1 + y2) / 4}
+			project.TransformEx(lonq, latq, make([]float64, 2), nil)
+			return computeDeltaOffset(lon1, lat1, lonm, latm, lonq[0], latq[0], x1, y1, (x1+x2)/2, (y1+y2)/2, project) + computeDeltaOffset(lonm, latm, lon2, lat2, lonq[1], latq[1], (x1+x2)/2, (y1+y2)/2, x2, y2, project)
+		}
+		return 360 * math.Round((lon1-lon2+2*deltaLon)/360)
+	}
+	return 0
+}
+
 func relativeAccuracy(project *godal.Transform, x, y, lon, lat []float64) []float64 {
 	// Create the midpoints
 	n := len(x) - 1
@@ -344,8 +363,11 @@ func relativeAccuracy(project *godal.Transform, x, y, lon, lat []float64) []floa
 	// Transform them to lon/lat
 	project.TransformEx(lonm, latm, make([]float64, len(lon)), nil)
 	// Estimate accuracy as a percentage of the length of each edge
-	// MidPoint are used to be sure that the geodetic line is the right one (eg when points do not belong to the same hemisphere)
+	// MidPoint are also used to be sure that the geodetic line is the right one (eg when points do not belong to the same hemisphere)
+	offsetLon := 0.0
 	for i := 0; i < n; i++ {
+		offsetLon += computeDeltaOffset(lon[i], lat[i], lon[i+1]+offsetLon, lat[i+1], lonm[i]+offsetLon, latm[i], x[i], y[i], x[i+1], y[i+1], project)
+		lon[i+1] += offsetLon
 		acc[i] = (lonLatDistance(lon[i], lat[i], lonm[i], latm[i]) + lonLatDistance(lon[i+1], lat[i+1], lonm[i], latm[i])) * accuracyPc
 	}
 	return acc
@@ -364,9 +386,11 @@ func (r Ring) cloneTo4326(crs *godal.SpatialRef, geodetic bool) (*Ring, error) {
 
 	// Project the shape in lon/lat
 	lon, lat := FlatCoordToXY(r.FlatCoords())
-	crsToLonLat.TransformEx(lon, lat, make([]float64, len(lon)), nil)
+	if err = crsToLonLat.TransformEx(lon, lat, nil, nil); err != nil {
+		return nil, fmt.Errorf("cloneTo4326.crsToLonLat: %w", err)
+	}
 
-	// Estimate accuracy
+	// Estimate accuracy and fix the dateline crossing
 	x, y := FlatCoordToXY(r.FlatCoords())
 	accuracyInMeter := relativeAccuracy(crsToLonLat, x, y, lon, lat)
 
@@ -409,14 +433,30 @@ func lonLatMidPoint(lon1, lat1, lon2, lat2 float64, geodetic bool) (float64, flo
 	}
 	lon1, lat1, lon2, lat2 = DegToRad*lon1, DegToRad*lat1, DegToRad*lon2, DegToRad*lat2
 	dlon := lon2 - lon1
-	bx := math.Cos(lat2) * math.Cos(dlon)
-	by := math.Cos(lat2) * math.Sin(dlon)
-	latm := math.Atan2(
-		math.Sin(lat1)+math.Sin(lat2),
-		math.Sqrt(square(math.Cos(lat1)+bx)+square(by)))
-	lonm := lon1 + math.Atan2(by, (math.Cos(lat1)+bx))
+	x1, y1, z1 := math.Cos(lat1), 0.0, math.Sin(lat1)
+	x2, y2, z2 := math.Cos(lat2)*math.Cos(dlon), math.Cos(lat2)*math.Sin(dlon), math.Sin(lat2)
+	x, y, z := (x1+x2)/2, (y1+y2)/2, (z1+z2)/2
+	if math.Abs(lon2-lon1) > math.Pi {
+		x, y, z = -x, -y, -z
+	}
+	latm := math.Atan2(z, math.Sqrt(square(x)+square(y)))
+	lonm := lon1 + math.Atan2(y, x)
 	return RadToDeg * lonm, RadToDeg * latm
 }
+
+// intermLatPointForLon returns the intermediate latitude of two lon/lat points in geographic coordinates (following the geodetic line) at longitude lonm
+/*func intermLatPointForLon(lon1, lat1, lon2, lat2, lonm float64) float64 {
+	lon1, lat1, lon2, lat2 = DegToRad*lon1, DegToRad*lat1, DegToRad*lon2, DegToRad*lat2
+	dlon := lon2 - lon1
+	x1, y1, z1 := math.Cos(lat1), 0.0, math.Sin(lat1)
+	x2, y2, z2 := math.Cos(lat2)*math.Cos(dlon), math.Cos(lat2)*math.Sin(dlon), math.Sin(lat2)
+	//find f such as lonm == lon1 + math.Atan2(f*y2, (1-f)*x1+f*x2)
+	tanM := math.Tan(lonm - lon1)
+	f := tanM * x1 / (y2 + tanM*(x1-x2))
+	x, y, z := (1-f)*x1+f*x2, (1-f)*y1+f*y2, (1-f)*z1+f*z2
+	latm := math.Atan2(z, math.Sqrt(square(x)+square(y)))
+	return RadToDeg * latm
+}*/
 
 type projection func(float64, float64) (float64, float64)
 
@@ -444,6 +484,9 @@ func densifyEdge(projToLonLat *projection, geodetic bool, x1, y1, x2, y2, lon1, 
 	xm, ym := (x1+x2)/2, (y1+y2)/2
 	lonm, latm := (*projToLonLat)(xm, ym)
 	lonm2, latm2 := lonLatMidPoint(lon1, lat1, lon2, lat2, geodetic)
+	if math.Abs(lon1-lon2) < 180 && math.Abs(lonm2-lonm) > 180 { // Handle the case when lon1 or lon2 does not belong to [-180, 180]
+		lonm = lonm2 + math.Mod(lonm-lonm2+540, 360) - 180
+	}
 
 	// Check whether the difference is lower than accuracy
 	distance := lonLatDistance(lonm, latm, lonm2, latm2)
@@ -463,3 +506,37 @@ func densifyEdge(projToLonLat *projection, geodetic bool, x1, y1, x2, y2, lon1, 
 		lonm, latm),
 		densifyEdge(projToLonLat, geodetic, xm, ym, x2, y2, lonm, latm, lon2, lat2, accuracy, recursion-1)...)
 }
+
+/*
+func clockWise(coords [][2]float64) bool {
+	// Find the bottom right point (can't be the last, because coords[0] == coords[end])
+	mini := 0
+	for i, c := range coords {
+		if c[0] < coords[mini][0] || (c[0] == coords[mini][0] && c[1] < coords[mini][1]) {
+			mini = i
+		}
+	}
+	// Compute the cross product
+	previ := mini - 1
+	if previ < 0 {
+		previ = len(coords) - 1
+	}
+	return (coords[mini+1][0]-coords[mini][0])*(coords[previ][1]-coords[mini][1]) < (coords[mini+1][1]-coords[mini][1])*(coords[previ][0]-coords[mini][0])
+}
+
+func clockWiseXY(x, y []float64) bool {
+	// Find the bottom right point (can't be the last, because coords[0] == coords[end])
+	mini := 0
+	for i := range x {
+		if x[i] < x[mini] || (x[i] == x[mini] && y[i] < y[mini]) {
+			mini = i
+		}
+	}
+	// Compute the cross product
+	previ := mini - 1
+	if previ < 0 {
+		previ = len(x) - 1
+	}
+	return (x[mini+1]-x[mini])*(y[previ]-y[mini]) < (y[mini+1]-y[mini])*(x[previ]-x[mini])
+}
+*/